requirements.txt
asgiref @ file:///C:/Users/BUILDE~1/AppData/Local/Temp/abs_cchykh5sgr/croots/recipe/asgiref_1661287542702/work
backports.zoneinfo @ file:///C:/Users/BUILDE~1/AppData/Local/Temp/abs_29zz4uumbw/croots/recipe/backports.zoneinfo_1661287541232/work
certifi @ file:///C:/b/abs_ac29jvt43w/croot/certifi_1665076682579/work/certifi
cffi @ file:///D:/bld/cffi_1656782950753/work
charset-normalizer==2.1.1
coreapi==2.3.3
coreschema==0.0.4
cryptography==38.0.1
defusedxml==0.7.1
Django @ file:///C:/Users/BUILDE~1/AppData/Local/Temp/abs_eeor5o9jj8/croots/recipe/django_1661292534510/work
django-templated-mail==1.1.1
djangorestframework @ file:///home/conda/feedstock_root/build_artifacts/djangorestframework_1663854154474/work
djangorestframework-jsonapi==6.0.0
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.2.1
djoser==2.1.0
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.1
oauthlib==3.2.1
path @ file:///opt/conda/conda-bld/path_1641578212155/work
Pillow==9.2.0
psycopg2==2.9.3
pycparser @ file:///home/conda/feedstock_root/build_artifacts/pycparser_1636257122734/work
PyJWT==1.7.1
python3-openid==3.2.0
pytz @ file:///home/conda/feedstock_root/build_artifacts/pytz_1664798238822/work
requests==2.28.1
requests-oauthlib==1.3.1
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.3.0
sqlparse @ file:///tmp/build/80754af9/sqlparse_1602184451250/work
tzdata @ file:///tmp/build/80754af9/python-tzdata_1611693927741/work
uritemplate==4.1.1
urllib3==1.26.12
wincertstore==0.2
